changes in test cases for Footer and Header
